+++ src/components/modals/RewardsModal.tsx
@@ -1,0 +1,91 @@
+import { useEffect, useState } from 'react'
+import { SparklesIcon } from '@heroicons/react/outline'
+import { useTranslation } from 'react-i18next'
+import { AchievementList } from '../achievements/AchievementList'
+import { CosmeticsPanel } from '../rewards/CosmeticsPanel'
+import { BaseModal } from './BaseModal'
+
+type RewardsTab = 'achievements' | 'cosmetics'
+
+type Props = {
+  isOpen: boolean
+  handleClose: () => void
+  isUppercase: boolean
+  excludeUrl: boolean
+  initialTab?: RewardsTab
+  scrollToAchievement?: string
+  onOpenDeadEndHelp?: () => void
+}
+
+export const RewardsModal = ({
+  isOpen,
+  handleClose,
+  isUppercase,
+  excludeUrl,
+  initialTab,
+  scrollToAchievement,
+  onOpenDeadEndHelp,
+}: Props) => {
+  const { t } = useTranslation()
+  const [activeTab, setActiveTab] = useState<RewardsTab>(
+    initialTab || 'achievements'
+  )
+  const [focusedAchievement, setFocusedAchievement] = useState<
+    string | undefined
+  >(scrollToAchievement)
+
+  useEffect(() => {
+    if (!isOpen) return
+    setActiveTab(initialTab || 'achievements')
+    setFocusedAchievement(scrollToAchievement)
+  }, [isOpen, initialTab, scrollToAchievement])
+
+  const tabs = [
+    { id: 'achievements' as const, label: t('achievements') },
+    { id: 'cosmetics' as const, label: t('cosmetics') },
+  ]
+
+  return (
+    <BaseModal
+      title={t('rewards')}
+      icon={<SparklesIcon />}
+      isOpen={isOpen}
+      handleClose={handleClose}
+    >
+      <div className="flex border-b border-gray-200 mb-4">
+        {tabs.map((tab) => (
+          <button
+            key={tab.id}
+            className={`px-4 py-2 text-sm font-medium ${
+              activeTab === tab.id
+                ? 'border-b-2 border-indigo-600 text-indigo-600 font-bold'
+                : 'text-gray-400 hover:text-gray-600'
+            }`}
+            onClick={() => setActiveTab(tab.id)}
+          >
+            {tab.label}
+          </button>
+        ))}
+      </div>
+
+      <div className="h-[26rem]">
+        {activeTab === 'achievements' && (
+          <AchievementList
+            scrollToId={focusedAchievement}
+            onOpenDeadEndHelp={onOpenDeadEndHelp}
+          />
+        )}
+        {activeTab === 'cosmetics' && (
+          <CosmeticsPanel
+            isUppercase={isUppercase}
+            excludeUrl={excludeUrl}
+            onNavigateToAchievement={(achievementId) => {
+              setFocusedAchievement(achievementId)
+              setActiveTab('achievements')
+            }}
+          />
+        )}
+      </div>
+    </BaseModal>
+  )
+}
