+++ src/locales/de/translation.json
@@ -22,6 +22,7 @@
   "deadEndRuleLink": "Dead-End-Regel",
   "guessDistribution": "Versuchsverteilung",
   "records": "Rekorde",
+  "rewards": "Belohnungen",
   "statistics": "Statistik",
   "totalTries": "Versuche gesamt",
   "successRate": "Erfolgsquote",
