+++ src/locales/en/translation.json
@@ -22,6 +22,7 @@
   "deadEndRuleLink": "Dead End rule",
   "guessDistribution": "Guess Distribution",
   "records": "Records",
+  "rewards": "Rewards",
   "statistics": "Statistics",
   "totalTries": "Total tries",
   "successRate": "Success rate",
