+++ src/locales/es/translation.json
@@ -19,6 +19,7 @@
   "notInWordInstructions": "La letra O no se encuentra en la palabra.",
   "guessDistribution": "Distribución de aciertos",
   "records": "Registros",
+  "rewards": "Recompensas",
   "statistics": "Estadísticas",
   "totalTries": "Jugadas",
   "successRate": "Victorias",
