+++ src/locales/ja/translation.json
@@ -22,6 +22,7 @@
   "deadEndRuleLink": "Dead End ルール",
   "guessDistribution": "推測の分布",
   "records": "記録",
+  "rewards": "報酬",
   "statistics": "統計",
   "totalTries": "総ゲーム数",
   "successRate": "成功率",
