+++ src/locales/ko/translation.json
@@ -22,6 +22,7 @@
   "deadEndRuleLink": "Dead End 규칙",
   "guessDistribution": "추측 분포",
   "records": "기록",
+  "rewards": "보상",
   "statistics": "통계",
   "totalTries": "총 게임 수",
   "successRate": "성공률",
