+++ src/locales/sw/translation.json
@@ -19,6 +19,7 @@
   "notInWordInstructions": "Herufi U haiko kwa neno, pahali yoyote.",
   "guessDistribution": "Usambazaji wa kubahatisha",
   "records": "Rekodi",
+  "rewards": "Zawadi",
   "statistics": "Takwimu",
   "totalTries": "Mara ya kujaribu",
   "successRate": "Kiwango ya kushinda",
