+++ src/locales/zh/translation.json
@@ -19,6 +19,7 @@
   "notInWordInstructions": "字母 U 不在正解當中。",
   "guessDistribution": "猜詞分佈",
   "records": "記錄",
+  "rewards": "奖励",
   "statistics": "統計",
   "totalTries": "總猜測次數",
   "successRate": "成功率",
